+++ src/typings/editor.ts
@@ -1,0 +1,3 @@
+export type Editors = 'vscode'
+
+export type EditorOptions = { type: Editors }